InputCurve
Renamed Utitiliy.java to Util.java. Added a line that creates an object from the util file so we can use methods within Util in the main robot class. No idea how that works. Also added an input curve function in util that takes an input from 0-1 and returns a curved version based on the linearity provided. Going to make a clamp function next.